Do not require Content-Length on responses that do not have a body. This was actually an existing requirement in RFC2616:
   Applications SHOULD use this field to indicate the transfer-length of
   the message-body, unless this is prohibited by the rules in section
   4.4.

because 4.4 does not prohibit sending of content-length, even though
it is believed what was meant by 2616 is that the SHOULD does not
apply to messages that prohibit a body.

Anyway, this change should clarify it, and the other requirements
on sending content-length, once and for all, at the cost of a bit
of duplication between p1 and p2.